Tidy: tidy old code and remove dead code
